Dissect the forms.py
