+++ accounts/forms.py
@@ -6,3 +6,16 @@ class LoginForm(forms.ModelForm):
     class Meta:
         model = User
         fields = ['username', 'password']
+
+class GenericUserForm(forms.ModelForm):
+    class Meta:
+        model = User
+        fields = ('username', 'password')
+
+class UpdateUserForm(GenericUserForm):
+    username = forms.CharField(max_length=150, label='New Username')
+    password = forms.CharField(label='New Password', widget=forms.PasswordInput)
+
+class CreateUserForm(GenericUserForm):
+    username = forms.CharField(max_length=150, label='Username')
+    password = forms.CharField(label='Password', widget=forms.PasswordInput)
