+++ cosmos/views.py
@@ -10,12 +10,9 @@
 from django.contrib.auth.decorators import login_required
 from django.http import HttpResponse
 from django.shortcuts import render, redirect, get_object_or_404
-from django.views.decorators.csrf import csrf_exempt
 
-from CS416FinalProject.forms import CreatePostForm
+from .forms import CreatePostForm
 from .models import Post, Reaction, User
-
-
 
 @login_required(login_url='login')
 def homePage(request):
@@ -78,6 +75,7 @@ def createReply(request, other_post=None):
     return render(request, 'cosmos/create-post.html', context)
 
 # the worst function in the entire project
+@login_required(login_url='login')
 def reactionAJAXOperations(request):
     def likeReturn(postID):
         return HttpResponse(Reaction.objects.filter(reaction='LIKE', post_id=postID).count())
@@ -86,8 +84,9 @@ def reactionAJAXOperations(request):
         return HttpResponse(Reaction.objects.filter(reaction='DISLIKE', post_id=postID).count())
 
     if request.method == 'GET':
-        return None
+        return redirect('landing')
 
+    # setter db operations
     if request.POST['operation'] == 'SET':
         # get the reaction associated with this post and user
         react = Reaction.objects.filter(reaction=request.POST['goal'], post_id=request.POST['postID'], user_id=request.POST['userID'])
@@ -100,6 +99,7 @@ def reactionAJAXOperations(request):
             # yes, delete it
             react[0].delete()
 
+    #getter db operations
     if request.POST['goal'] == 'LIKE':
         return likeReturn(request.POST['postID'])
